SnowDown/requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-cors-headers==3.10.0
django-environ==0.4.5
django-graphql-jwt==0.3.4
django-storages==1.11.1
dnspython==2.1.0
google-api-core==1.31.4
google-auth==1.30.0
google-cloud-core==2.1.0
google-cloud-secret-manager==2.4.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpc-google-iam-v1==0.12.3
grpcio==1.41.1
gunicorn==20.1.0
idna==3.3
libcst==0.3.21
mypy-extensions==0.4.3
packaging==21.2
Pillow==8.4.0
promise==2.3
proto-plus==1.19.7
protobuf==3.19.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pymongo==3.12.1
pyparsing==2.4.7
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.2.4
text-unidecode==1.3
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
